backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
pydantic>=2.5.2
mistralai==0.0.7
python-jose==3.3.0
werkzeug==2.3.7
gunicorn==21.2.0